section2_final_2024-06-11

+++ src/main/java/com/example/entity/BookDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookDTO {
-    private Long num;
+    private int num;
     private String title;
     private int price;
     private String author;
